% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{example}
\format{
A object with 30 rows and 34 variable:
\describe{
\item{example}{annual max daily rainfall in mm}
}
}
\usage{
example
}
\description{
A data frame with annual max daily rainfall series
}
\keyword{datasets}
